Add option of 3 phases/ternary with quadratic energies

+++ tests/Sintering/test2d.py
@@ -3,25 +3,24 @@ import sys
 import subprocess
 import os
 
-print("Test One grains with multiple order parameters...")
+print("Test Two grains sintering...")
 
 mpicmd  = sys.argv[1]+" "+sys.argv[2]+" "+sys.argv[3]
 exe     = sys.argv[4]
 inp     = sys.argv[5]
 datadir = sys.argv[6]
 
-#make symbolic link to input data
-data = "spheres.csv"
+#make symbolic link to calphad data
+data = "2spheres.csv"
 if not os.path.exists(data):
   src = datadir+'/'+data
-  print("Create symlink {}".format(src))
   os.symlink(src, data)
 
 #prepare initial conditions file
-initfilename="sphere.nc"
+initfilename="2spheres.nc"
 subprocess.call(["python3", "../../utils/make_multi_spheres.py",
   "--nx", "64", "--ny", "64", "--nz", "1",
-  "--concentration-A", "0.1", "--concentration-out", "0.06",
+  "--concentration-A", "1.,0.", "--concentration-B", "0.,1.", "--concentration-out", "0.,0.",
   "--spheres", data,
   initfilename])
 
@@ -31,11 +30,15 @@ output = subprocess.check_output(command,shell=True)
 
 #analyse AMPE standard output
 lines=output.split(b'\n')
-volumes=[]
+volfractions=[]
 
 end_reached = False
-target_sf = 0.19
 for line in lines:
+  if line.count(b'phase') and line.count(b'Volume'):
+    print(line)
+    words=line.split()
+    volume=eval(words[6])
+    volfractions.append(volume)
 
   if line.count(b'cycle'):
     print(line)
@@ -44,15 +47,24 @@ for line in lines:
     if time>0.25:
       end_reached = True
 
-  if line.count(b'fraction')  and line.count(b'phase 0'):
-    print(line)
-    if end_reached:
-      words=line.split()
-      sfraction=eval(words[6])
-      if abs(sfraction-target_sf)>1.e-2:
-        print("Wrong solid fraction:")
-        print("found {}, expected {}".format(sfraction-target_sf, sfraction-target_sf))
-        sys.exit(1)
+minv=1.
+maxv=0.
+for v in volfractions:
+  if v<minv:
+    minv = v
+  if v>maxv:
+    maxv = v
+
+expected_value=0.236
+if abs(maxv-expected_value)>0.001:
+  print("Expected maxv = {}, found {}".format(expected_value,maxv))
+  sys.exit(1)
+
+expected_value=0.037
+if abs(minv-expected_value)>0.001:
+  print("Expected minv = {}, found {}".format(expected_value,minv))
+  sys.exit(1)
+
 
 os.remove(initfilename)
 os.unlink(data)
@@ -60,5 +72,6 @@ os.unlink(data)
 if end_reached:
   sys.exit(0)
 else:
-  print("End time not reached")
+  print("End time not reached!")
   sys.exit(1)
+
